Update all quotas schema tests
