+++ tests/schemas/test_network_quota_invalid_schema.py
@@ -1,0 +1,28 @@
+from typing import Any
+
+import pytest
+from pytest_cases import parametrize_with_cases
+
+from fed_reg.provider.schemas_extended import NetworkQuotaCreateExtended
+from fed_reg.quota.models import NetworkQuota
+from fed_reg.quota.schemas import NetworkQuotaBase, NetworkQuotaRead
+
+
+@parametrize_with_cases("key, value")
+def test_invalid_base(key: str, value: Any) -> None:
+    d = {key: value}
+    with pytest.raises(ValueError):
+        NetworkQuotaBase(**d)
+
+
+@parametrize_with_cases("key, value")
+def test_invalid_read(network_quota_model: NetworkQuota, key: str, value: str) -> None:
+    network_quota_model.__setattr__(key, value)
+    with pytest.raises(ValueError):
+        NetworkQuotaRead.from_orm(network_quota_model)
+
+
+def test_invalid_create_extended() -> None:
+    d = {}
+    with pytest.raises(ValueError):
+        NetworkQuotaCreateExtended(**d)
