+++ tests/schemas/test_quota_schema.py
@@ -2,7 +2,6 @@ from typing import Literal
 
 from pytest_cases import parametrize, parametrize_with_cases
 
-from fed_reg.models import BaseNode
 from fed_reg.quota.schemas import QuotaBase
 from tests.utils import random_lower_string
 
@@ -14,14 +13,13 @@ class CaseAttr:
     def case_desc(self) -> tuple[Literal["description"], str]:
         return "description", random_lower_string()
 
-    @parametrize(value=[True, False])
+    @parametrize(value=(True, False))
     def case_boolean(self, value: bool) -> tuple[Literal["per_user"], bool]:
         return "per_user", value
 
 
 @parametrize_with_cases("key, value", cases=CaseAttr)
 def test_base_public(key: str, value: str) -> None:
-    assert issubclass(QuotaBase, BaseNode)
     d = {key: value} if key else {}
     item = QuotaBase(**d)
     assert item.description == d.get("description", "")
